+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ## v0.4.8 - 2020
 
 - Added other handlings of user notifications. prompt_user_loop, tell_user and tell_user_then_kill. prompt_user_loop is not exiting on the not now button, but is waiting an hour until next question, and after that it changes action to tell_user. tell_user is just a dialog to the user to tell that an update is needed, with no way of dismissing, but they can leave the dialog open. it's more eager to install, and do not allow the user to cancel the upgrade. if _then_kill is added in the name, it will ask 2 times and kill it the 3. What could happen is that after the polite quite, the user could open the app, and if so, he/she will be asked again. The workings of these have been reworked a bit. If the app is killed the wait for that is only 5 seconds.
-- Labels: improved 8x8
+- Labels improved: 8x8
+- Labels added with versioncheck: iterm2, docker, postman, sourcetree, visualstudiocode, malwarebytes, suspiciouspackage
 
 
 ## v0.4.7 - 2020-10-26